Added graphical map when looking at map
I'm not sure how that is supposed to work,
for now I just make the client ask for
every instance of an image. However this
may be the wrong way to do this. I might
need to send an actual image that I generate
offline, and just set a couple of clickable
areas. That would allow me to do multiple
layers, like add towns to the landscape, which
is not possible as of yet.